updated to work with transactions

+++ conductor/conductor.go
@@ -136,16 +136,16 @@ func debugF(format string, args ...interface{}) {
 }
 
 func newBlock() (upper, lower []byte, blockheight uint32) {
-	blockHeight := 433789
-	blockFees := 8756123 // satoshi
+	blockHeight := uint32(433789)
+	blockFees := 8756123 // satoshis
 	pubkey := "0225c141d69b74adac8ab984a8eb9fee42c4ce79cf6cb2be166b1ddc0356b37086"
-	// conductor generates this ..
+	// conductor generates this ...
 	upper, lower, err := coin.CoinbaseTemplates(blockHeight, blockFees, pubkey)
 	if err != nil {
 		log.Fatalf("failed to generate coinbase: %v", err)
 	}
 	// sends upper, lower , blockHeight --> server
-	return upper, lower, uint32(blockHeight)
+	return upper, lower, blockHeight
 }
 
 func main() {
